@@ -1,64 +1,62 @@
-﻿using Microsoft.Extensions.Caching.Memory;
-using System;
+﻿using System;
 using System.Collections.Concurrent;
 using System.Linq.Expressions;
-using System.Reflection;
+using Microsoft.Extensions.Caching.Memory;
 
-namespace PersistentMemoryCache.Internal
+namespace PersistentMemoryCache.Internal;
+
+public abstract class LiteDbCacheEntry
 {
-    public abstract class LiteDbCacheEntry
+    private static ConcurrentDictionary<Type, Delegate> _constructors = new();
+    private static Type _liteDbCacheEntryOpenType = typeof(LiteDbCacheEntry<>);
+    private static Type[] _emptyTypesArray = [];
+
+    private TimeSpan? _slidingExpiration;
+
+    public int Id { get; set; }
+    public string CacheName { get; set; }
+    public CacheItemPriority Priority { get; set; }
+    public object Key { get; set; }
+    internal DateTimeOffset LastAccessed { get; set; }
+    public DateTimeOffset? AbsoluteExpiration { get; set; }
+    public TimeSpan? SlidingExpiration
     {
-        private static ConcurrentDictionary<Type, Delegate> _Constructors = new ConcurrentDictionary<Type, Delegate>();
-        private static Type _LiteDbCacheEntryOpenType = typeof(LiteDbCacheEntry<>);
-        private static Type[] _EmptyTypesArray = new Type[0];
-
-        private TimeSpan? _SlidingExpiration;
-
-        public int Id { get; set; }
-        public string CacheName { get; set; }
-        public CacheItemPriority Priority { get; set; }
-        public object Key { get; set; }
-        internal DateTimeOffset LastAccessed { get; set; }
-        public DateTimeOffset? AbsoluteExpiration { get; set; }
-        public TimeSpan? SlidingExpiration
+        get
         {
-            get
+            if (_slidingExpiration <= TimeSpan.Zero)
             {
-                if (_SlidingExpiration <= TimeSpan.Zero)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(SlidingExpiration), _SlidingExpiration, "The sliding expiration value must be positive.");
-                }
-                return _SlidingExpiration;
+                throw new ArgumentOutOfRangeException(nameof(SlidingExpiration), _slidingExpiration, "The sliding expiration value must be positive.");
             }
-            set
-            {
-                if (value <= TimeSpan.Zero)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(SlidingExpiration), value, "The sliding expiration value must be positive.");
-                }
-                _SlidingExpiration = value;
-            }
+            return _slidingExpiration;
         }
-
-        public static LiteDbCacheEntry ConstructCacheEntry(Type type) => (LiteDbCacheEntry)_Constructors.GetOrAdd(type, cacheType =>
+        set
         {
-            Type cacheEntryClosedType = _LiteDbCacheEntryOpenType.MakeGenericType(type);
-            ConstructorInfo constructor = cacheEntryClosedType.GetConstructor(_EmptyTypesArray);
-            Type delegateType = typeof(Func<>).MakeGenericType(cacheEntryClosedType);
-            LambdaExpression lambda = Expression.Lambda(delegateType, Expression.New(constructor));
-            return lambda.Compile();
-        }).DynamicInvoke();
-        
-        public abstract object GetValue();
-        public abstract void SetValue(object value);
+            if (value <= TimeSpan.Zero)
+            {
+                throw new ArgumentOutOfRangeException(nameof(SlidingExpiration), value, "The sliding expiration value must be positive.");
+            }
+            _slidingExpiration = value;
+        }
     }
 
-    public class LiteDbCacheEntry<T> : LiteDbCacheEntry
+    public static LiteDbCacheEntry ConstructCacheEntry(Type type) => (LiteDbCacheEntry)_constructors.GetOrAdd(type, cacheType =>
     {
-        public T Value { get; set; }
+        var cacheEntryClosedType = _liteDbCacheEntryOpenType.MakeGenericType(type);
+        var constructor = cacheEntryClosedType.GetConstructor(_emptyTypesArray);
+        var delegateType = typeof(Func<>).MakeGenericType(cacheEntryClosedType);
+        var lambda = Expression.Lambda(delegateType, Expression.New(constructor));
+        return lambda.Compile();
+    }).DynamicInvoke();
 
-        public override object GetValue() => Value;
+    public abstract object GetValue();
+    public abstract void SetValue(object value);
+}
 
-        public override void SetValue(object value) => Value = (T)value;
-    }
+public class LiteDbCacheEntry<T> : LiteDbCacheEntry
+{
+    public T Value { get; set; }
+
+    public override object GetValue() => Value;
+
+    public override void SetValue(object value) => Value = (T)value;
 }
