@@ -1,18 +1,13 @@
 ﻿using System;
 
-namespace PersistentMemoryCache
-{
-    public class LiteDbOptions
-    {
-        public LiteDbOptions(string fileName)
-        {
-            if (fileName == null)
-            {
-                throw new ArgumentNullException(nameof(fileName));
-            }
-            FileName = fileName;
-        }
+namespace PersistentMemoryCache;
 
-        public string FileName { get; }
+public class LiteDbOptions
+{
+    public LiteDbOptions(string fileName)
+    {
+        FileName = fileName ?? throw new ArgumentNullException(nameof(fileName));
     }
+
+    public string FileName { get; }
 }
